archive/requirements.txt
pylint
pytest
pymssql
python-dotenv